==> analysis/scripts/05-combine-asvs-and-fluxes.R <==

library(tidyverse)
library(decontam)
library(knitr)
library(vegan)
source('R/functions.R')

load(file = "analysis/hpc_output/dada2_output.RData")

samples <- read_tsv(paste0(der_dir, "barcode_master.tsv"))
fluxes <- read_tsv(paste0(der_dir, "fluxes.tsv"))
qubit <- read_tsv(paste0(raw_dir, "post_pcr_qubit.tsv"))


metadata <-
  samples %>% 
  rename(sample = sample_id) %>% 
  left_join(qubit, by = c("pcr_well" = "well")) %>% 
  mutate(treat = (Vectorize(function (x) {if(x %in% c('PCRP', 'PCRN')) x
    else substr(x, 1, 3)}))(sample)) %>% 
  mutate(col = as.character(col)) %>% 
  select(sample, row, col, treat, qubit_ngml)

rm(qubit)
rm(samples)

# Check if taxa names match between the otu and taxa tables
sum(colnames(seqtab.nochim) != rownames(taxa))
# They match if = 0 
seqs <- tibble(asv = paste0("ASV", seq(ncol(seqtab.nochim))),
               seq = colnames(seqtab.nochim))
seq_out <- seqs
names(seq_out) <- c('seq.name', 'seq.text')
phylotools::dat2fasta(seq_out, 'analysis/output/refseqs.fna')

colnames(seqtab.nochim) <- paste0("ASV", seq(ncol(seqtab.nochim)))
rownames(seqtab.nochim) <- metadata$sample
rownames(taxa) <- colnames(seqtab.nochim)

asvs <-
  seqtab.nochim %>% 
  as_tibble(rownames = "sample")

asvs %>% 
    pivot_longer(-sample, names_to = 'asv', values_to = 'count') %>% 
    pivot_wider(names_from = sample, values_from = count)%>% 
    as.data.frame ->
    asv_to_biom

rownames(asv_to_biom) <- asv_to_biom[,1]
asv_to_biom <- asv_to_biom[,-1]
asv_biom <- biomformat::make_biom(asv_to_biom)
biomformat::write_biom(asv_biom, 'analysis/output/asv.biom')

rm(seqtab.nochim)

taxa <-
  taxa %>% 
  as_tibble(rownames = "asv")

aes_data <-
  asvs %>% 
  pivot_longer(-sample, names_to = "asv", values_to = "count") %>% 
  inner_join(metadata, by = "sample") %>% 
  inner_join(taxa, by = "asv")


# Alpha Diversity for All Data

# Here is alpha diversity for each treatment, positive/negative control, etc.
# In addition, there is alpha diversity by row and column of the PCR plate. I don't see any relationship between PCR plate location and alpha diversity, which is good. It looks like the PCR positive control, which was a pure culture of E. Coli, has particularly low diversity, which is good. Let's pull that sample out and look at it.


aes_data %>%
  select(sample, treat, asv, count) %>% 
  filter(count > 0) %>% 
  count(sample, treat, count) %>%
  nest(data = -c(sample,treat)) %>%
  mutate(n_seqs = map_dbl(data, ~sum(.x$count * .x$n)),
         sobs = map_dbl(data, ~sum(.x$n))) %>%
  select(-data) %>%
  select(-sample) %>% 
  group_by(treat) %>% 
  rename(Reads = n_seqs, Richness = sobs) %>% 
  pivot_longer(-c(treat)) %>%
  ggplot(aes(x=treat, y=value)) +
  facet_wrap(~ name, nrow = 3, scales = "free") +
  geom_boxplot() +
  geom_jitter() +
  theme_classic() +
  labs(x = "Treatment", y = "Value")


aes_data %>%
  select(row, col, asv, count) %>% 
  filter(count > 0) %>% 
  mutate(row = as.numeric(factor(row)),
         col = as.numeric(col)) %>% 
  count(row, col, count) %>%
  nest(data = -c(row, col)) %>%
  mutate(n_seqs = map_dbl(data, ~sum(.x$count * .x$n)),
         sobs = map_dbl(data, ~sum(.x$n))) %>%
  select(-data) %>%
  rename(Column = col, Row = row, Richness = sobs, Sequences = n_seqs) %>% 
  pivot_longer(c(Row, Column), names_to = "row_or_column", values_to = "position") %>%
  pivot_longer(c(Sequences, Richness), names_to = "name", values_to = "count") %>%
  ggplot(aes(x=position, y=count)) +
  facet_grid(name ~ row_or_column, scales = "free") +
  geom_point() +
  theme_classic() +
  geom_smooth() + 
  scale_x_continuous(breaks = 1:9)

## Top 20 taxa in controls
#
#top20_taxa_in_treatment <- function(x, treatment) {
#    print(paste0('Top 20 taxa in ', treatment))
#x %>% 
#  filter(treat == treatment) %>% 
#  select(-sample, -row, -col, -qubit_ngml) %>% 
#  group_by(asv) %>% 
#  mutate(count = sum(count)) %>%  
#  ungroup() %>% 
#  distinct() %>% 
#  arrange(desc(count)) %>% 
#  slice_head(n = 20)
#}
#
#### Potting Mix
#
#top20_taxa_in_treatment(aes_data, "PM0")
#
#
#### Negative Control
#
#top20_taxa_in_treatment(aes_data, "NC0")
#
#
#### PCR Positive Control
#
#top20_taxa_in_treatment(aes_data, "PCRP")
#
#
#### PCR Negative Control
#
#top20_taxa_in_treatment(aes_data, "PCRN")
#
#
#### Soil Core
#
#top20_taxa_in_treatment(aes_data, "SC0")

# Sequence Depth Distribution


# Look at distribution of sequencing depths
aes_data %>%
  select(sample, asv, count) %>% 
  group_by(sample) %>% 
  summarize(total = sum(count)) %>% 
  ggplot(aes(x = total)) +
  geom_histogram() +
  labs(x = "Sequences", y = "Count")


# Beta Diversity all samples

aes_df <- 
  aes_data %>%
  select(sample, asv, count) %>% 
  pivot_wider(names_from=asv, values_from=count, values_fill = 0) %>%
  as.data.frame()
rownames(aes_df) <- aes_df$sample
aes_df <- aes_df[,-1]

dist <- vegdist(aes_df, method="bray")
pcoa <- cmdscale(dist, eig = TRUE, add = TRUE)
positions = pcoa$points
colnames(positions) <- paste0("pcoa", 1:ncol(positions))

percent_explained <- 100 * pcoa$eig / sum(pcoa$eig)

tibble(pe = percent_explained, 
       axis=1:length(percent_explained)) %>% 
  ggplot(aes(x = axis, y = pe)) +
  geom_line() +
  coord_cartesian(xlim = c(1,10)) +
  labs(title = "Bray PCoA Scree Plot", y = "Precent Explained", x = "PCo Axis")

labs <- c(paste0("PCo 1 (", format(round(percent_explained[1], 1), nsmall = 1), "%)"),
          paste0("PCo 2 (", format(round(percent_explained[2], 1), nsmall = 1), "%)"))

positions <- 
  positions %>% 
  as_tibble(rownames = "sample")

bray_data <-
  aes_data %>% 
  select(sample, treat, row, col) %>% 
  distinct %>% 
  inner_join(positions, by = "sample")

bray_data %>% 
  ggplot(aes(x = pcoa1, y = pcoa2, color = treat)) +
  geom_point() +
  labs(x = labs[1], y = labs[2], title = "Bray PCoA by Treatment") +
  coord_fixed(percent_explained[2]/percent_explained[1], clip="off")


bray_data %>% 
  ggplot(aes(x = pcoa1, y = pcoa2, color = col)) +
  geom_point() +
  labs(x = labs[1], y = labs[2], title = "Bray PCoA by PCR plate column") +
  coord_fixed(percent_explained[2]/percent_explained[1], clip="off")

bray_data %>% 
  ggplot(aes(x = pcoa1, y = pcoa2, color = row)) +
  geom_point() +
  labs(x = labs[1], y = labs[2], title = "Bray PCoA by PCR plate row") +
  coord_fixed(percent_explained[2]/percent_explained[1], clip="off")


# DECONTAM


aes_data %>% 
  filter(!treat %in% c('PM0', 'SC0', 'PCRN', 'PCRP')) %>% 
  mutate(ctrl_smpl = case_when(treat == "NC0" ~ "control", 
                               TRUE ~ "sample")) %>% 
  add_count(sample, count) %>%
  nest(data = -c(sample, ctrl_smpl)) %>%
  mutate(n_seqs = map_dbl(data, ~sum(.x$count * .x$n))) %>% 
  arrange(n_seqs) %>% 
  mutate(index = seq(nrow(.))) %>% 
  ggplot(aes(x=index, y=n_seqs, color=ctrl_smpl)) + 
  geom_point() +
  labs(y = "Library Size", x = "Index")



# My negative controls have a lower library size than my samples.

## Decontam Combined Prevalence and Frequency

# Here, I use DNA concentration values from Qubit and prevalence in negative
# controls to identify potential contaminants. Several negative controls
# had DNA concentration below the detection limit of the instrument and these
# were set to 0.50 ng/mL, which is close to the DNA concentration of the
# negative controls that were within the detection limit and is at the limit of the 
# Qubit's detection limit.
# 
# In the plots below, ASV1 is a true sequence. The others are the ASVs identified
# by decontam as potential contaminants.


otu_table <-
  aes_data %>% 
  filter(!treat %in% c('PM0', 'SC0', 'PCRN', 'PCRP')) %>% 
    mutate(control = case_when(treat == "NC0" ~ TRUE, 
                               TRUE ~ FALSE)) %>% 
  select(qubit_ngml, control, sample, asv, count) %>% 
  pivot_wider(names_from = asv, values_from = count) %>% 
  mutate(qubit_ngml = case_when(qubit_ngml == 0 ~ 0.5, 
                               TRUE ~ qubit_ngml)) %>% 
  as.data.frame
concentration <- otu_table$qubit_ngml
otu_table <- otu_table[, -1]
control <- otu_table$control
otu_table <- otu_table[, -1]
rownames(otu_table) <- otu_table$sample
otu_table <- otu_table[, -1]
otu_mat <- as.matrix(otu_table)

contamdf <- isContaminant(otu_mat, conc = concentration, neg=control)

alleged_contaminants <- which(contamdf$contaminant)

quant_test <- 
  otu_mat[,c(1,alleged_contaminants)] %>% 
  as.data.frame %>% 
  as_tibble(rownames = "sample") %>% 
  mutate(conc = concentration) %>% 
  mutate(cont = control) %>% 
  pivot_longer(starts_with("ASV"))

ggplot(quant_test, aes(x = conc, y = value, color = cont)) +
  facet_wrap(~ name, scales = "free") +
  geom_point()

contaminants <- 
  which(contamdf$contaminant)[which(contamdf$contaminant) %in% 
  c(570, 810, 1283, 1859, 2233, 2777, 2926, 5687, 6696)]

save(contamdf, contaminants, file = paste0(der_dir, "decontam.RData"))
# 
# Decontam identified `r length(which(contamdf$contaminant))` potential contaminants based
# on prevalence and frequency. Visual inspection of abundance ~ concentration
# plots indicated that `r length(contaminants)` were likely contaminants. These
# ASVs were removed.

aes_data_no_contam <-
 aes_data %>%
  nest(data = -asv) %>% 
  filter(!asv %in% paste0("ASV", contaminants)) %>% 
  unnest(data)



# Combine ASV table and flux data



flux_sd <- 
  fluxes %>% 
  mutate(sample = paste0('G', passage, toupper(treat), rep)) %>% 
  select(sample, flux_date:selected) %>% 
  mutate(passage = as.character(passage)) %>% 
  select(sample, ch4, selected)

full_data <-
  aes_data %>% 
  filter(!treat %in% c("NC0",  "PM0",  "SC0",  "PCRP", "PCRN")) %>% 
  select(-c(row, col, qubit_ngml)) %>% 
  left_join(flux_sd, by = "sample")

# Remove zero-abundance ASVs
full_data <-
  full_data %>% 
  group_by(asv) %>% 
  mutate(total = sum(count)) %>% 
  filter(total != 0) %>% 
  ungroup() %>% 
  select(-total) %>% 
  filter(count > 0)

full_data <-
  full_data %>% 
  left_join(seqs, by = "asv")

write_tsv(full_data, paste0(der_dir, "community_data.tsv"))
